Create write_to_SD.cpp
Functions to initialize SD card which should be called in setup, and to write voltage data to SD card which should be called in measureVoltage function. Could be worthwhile to perhaps write to the SD card in read_register_groups so we can write "corrupted data" or some other string to know which data points are reliable.